Implement `workspace-users` API for `backend-go` (#491)
* Implement workspace users management API

* Add integration test

* Fix page size to total length

* Enhance error handing and sorting logic

* Refactor query and variable name

* Refactor test

* Use else case

+++ backend-go/api/codepair/v1/models/model_find_workspace_users_response.go
@@ -9,5 +9,5 @@ type FindWorkspaceUsersResponse struct {
 	Cursor string `json:"cursor"`
 
 	// Total count of workspace users
-	TotalLength float32 `json:"totalLength"`
+	TotalLength int64 `json:"totalLength"`
 }
